+++ source/MonoGame.Extended.Gui/Cursor.cs
@@ -5,7 +5,7 @@ namespace MonoGame.Extended.Gui
 {
     public class Cursor
     {
-        public TextureRegion2D TextureRegion { get; set; }
+        public TextureRegion TextureRegion { get; set; }
         public Color Color { get; set; } = Color.White;
     }
 }