+++ source/MonoGame.Extended.Gui/Serialization/GuiTextureRegionService.cs
@@ -8,7 +8,7 @@ namespace MonoGame.Extended.Gui.Serialization
     public interface IGuiTextureRegionService : ITextureRegionService
     {
         IList<TextureAtlas> TextureAtlases { get; }
-        IList<NinePatchRegion2D> NinePatches { get; }
+        IList<NinePatchRegion> NinePatches { get; }
     }
 
     public class GuiTextureRegionService : TextureRegionService, IGuiTextureRegionService