+++ source/MonoGame.Extended/Sprites/ISpriteBatchDrawable.cs
@@ -7,7 +7,7 @@ namespace MonoGame.Extended.Sprites
     public interface ISpriteBatchDrawable
     {
         bool IsVisible { get; }
-        TextureRegion2D TextureRegion { get; }
+        TextureRegion TextureRegion { get; }
         Vector2 Position { get; }
         float Rotation { get; }
         Vector2 Scale { get; }